[tlo-parsing] Rename tlo-parsing-tools.h -> tlo-parsing.h.
