+++ common/tlo-parsing/tlo-parser.cpp
@@ -12,7 +12,7 @@
 #include <unordered_map>
 
 #include "common/tl/tls.h"
-#include "common/tlo-parsing/tlo-parsing-tools.h"
+#include "common/tlo-parsing/tlo-parsing.h"
 
 namespace vk {
 namespace tl {
